Homework now allows fractional scores, e.g., 0.25 points on each question.

+++ src/main/java/org/chemvantage/HWTransaction.java
@@ -34,7 +34,7 @@ public class HWTransaction implements Serializable {
 	@Index	long topicId;
 	@Index	String userId;
 	@Index	Date graded;
-	@Index	int score;
+	@Index	double score;
 	@Index	long assignmentId;
 			int possibleScore;
 			String topicTitle;
@@ -44,7 +44,7 @@ public class HWTransaction implements Serializable {
 	
     HWTransaction() {}
     
-    HWTransaction(long questionId,String hashedId,Date graded,int score,long assignmentId,int possibleScore,String showWork) {
+	HWTransaction(long questionId,String hashedId,Date graded,double score,long assignmentId,int possibleScore,String showWork) {
     	this.questionId = questionId;
     	this.userId = hashedId;
         this.graded = graded;
